First working version of "set_gametile"
Console command to place gametiles in the current map.
Ground work for the server sending map patches to the client.

The current implementation is horrible. It is spawning a thread
which runs a python script that the user has to place in the maps folder

So this needs some work to be polished and portable.